Fix windows compilation
